+++ docs/CHANGELOG.md
@@ -5,8 +5,13 @@
 ### Core:
 - Package `tensors`
   - Improving support for sub-byte data types (`Int4`, `Int2`, `Uint4`, `Uint2`)
+  - Added `FromShapeForBackend()` to create new tensors with shared memory
+    if possible.
+  - Added `ToDevice()` and improved performance of MaterializeOnDevice when copying to shared buffers.
+  - Improved `MutableBytes()`, should improve the performance in some cases.
 
-### Under `pkg/ml`:
+### Packages in `pkg/ml`
+
 - Package `activations`
   - Added `HardSigmoid` activation.
   - Modified the parametrized activations to be suffixed with `With`: `LeakyReluWith`, `HardSigmoidWith`, `HardSwishWith`.
@@ -16,10 +21,17 @@
   - Dropout now takes a `*Node` instead of a float64, allowing dynamic dropout control.
 - Package `attention/pos`:
   - Split `Encoder` API into `QKEncoder` and `PreEncoder`, to support different types of positional encoders.
+  - Added `WithSlidingWindow` to support "sliding attention" (the slow way).
 - Package `transfromers`:
   - Updates and fixes to the API; Added methods to build partial models: `AllLayers`, `ForwardLayer`, `LogitsFromEmbeddings`, `EmbedTokesn`, etc.
   - Updated positional-encoder support.
-  - Added options `WithFinalNormalization`, `WithScalingOfTokenEmbeddings` and `WithArchitecture`.
+  - Added options `WithFinalNormalization`, `WithScalingOfTokenEmbeddings`, `WithArchitecture`, `WithSlidingWindow`
+    and `WithLayerTypes`.
+
+### Backends:
+
+- Package `xla`:
+  - `BufferToFlatData` and `BufferFromFlatData` now transfer using raw-bytes, for a slight gain in performance.
 
 # 0.27.2: DotGeneral with AccumulatorDType; Transformer architecture parameter;